requirements.txt
pandas==2.2.2
numpy==1.26.4
mlflow==2.16.0
pyyaml==6.0.2
tenacity==9.0.0
requests==2.32.3
scipy==1.13.1
